requirement.txt
aiogram==2.25.1
python-decouple==3.8
sqlalchemy==2.0.23